[quant] Add int8 per token dynamic quant + int4 per group quant for ExecuTorch
Summary:
att

Adding this for accuracy evaluation, we also added this in executorch repo and we'll dedup later

Test Plan:
python quantize.py --checkpoint_path checkpoints/$MODEL_REPO/model.pth --mode 8da4w-gptq --calibration_limit 5
python eval.py --checkpoint_path checkpoints/$MODEL_REPO/model_8da4w-gptq.g32.pth --tasks wikitext

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 864a12be8301b413af6a7073edefbc71b3a6049e
Pull Request resolved: https://github.com/pytorch-labs/gpt-fast/pull/102

+++ eval.py
@@ -220,7 +220,8 @@ def main(
     assert tokenizer_path.is_file(), tokenizer_path
 
     device = 'cuda'
-    precision = torch.bfloat16
+    # precision = torch.bfloat16
+    precision = torch.float16
 
     print("Loading model ...")
     t0 = time.time()
